requirements.txt
multilingual-clip
torch
transformers
numpy
scikit-learn
tqdm
Pillow
